pyproject.toml
[project]
name = "nonebot-plugin-ludokit"
version = "0.1.0"
description = "A NoneBot Plugin Library for Building Interactive Games."
authors = [
    { name = "KomoriDev", email = "mute231010@gmail.com" },
]
license = { text = "MIT" }
readme = "README.md"
keywords = []

requires-python = ">=3.10"
dependencies = [
    "nonebot-plugin-alconna>=0.57.4",
    "nonebot2>=2.4.2",
]

[project.urls]
homepage = "https://github.com/KomoriDev/nonebot-plugin-ludokit"
repository = "https://github.com/KomoriDev/nonebot-plugin-ludokit"

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot2[fastapi]>=2.4.2",
    "pre-commit>=4.2.0",
    "ruff>=0.11.3",
]

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]

[tool.ruff.lint.isort]
length-sort = true
force-sort-within-sections = true
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "standard"
